requirements.txt
PyJWT==2.4.0
bcrypt==4.0.1
pytest==7.4.0
Flask==2.0.1
Flask-PyMongo==2.3.0
python-dotenv==0.19.0
Werkzeug==2.0.3
pymongo==3.12.0
dnspython==2.3.0
gunicorn==20.1.0